Added cast to quiet MSVC for 32-bit compilation.
git-svn-id: svn+ssh://svn.code.sf.net/p/xsb/src/trunk@8876 756869e0-9e47-4bc5-b0a8-847305dc5e1c

+++ XSB/emu/tr_utils.c
@@ -5858,7 +5858,7 @@ int table_inspection_function( CTXTdecl ) {
     subgoal = (VariantSF)(tcp_subgoal_ptr(tcp));
     //    subgoal = (VariantSF) ptoc_int(CTXTc 2);
     //    printf("san %d\n",get_subgoal_answer_number(subgoal));
-    if ( subg_ans_ctr(subgoal) +1 >= ptoc_int(CTXTc 3) &&
+    if ( subg_ans_ctr(subgoal) +1 >= (unsigned)ptoc_int(CTXTc 3) &&
 	 !subg_is_completed(subgoal)) {
        schedule_ec(subgoal); 
        //       printf("scheduling ec\n");
